Collect common elements among legacy decoders
A small flourish adds support to decrypt Reduced elements.
This capability is yet untested, and unlikely to be needed for
any archive in the wild.
